+++ app/src/Routes.tsx
@@ -17,6 +17,7 @@ import { CloneProject } from "./pages/CloneProject";
 import { HistoryProject } from "./pages/HistoryProject";
 import { About } from "./pages/About";
 import { StatsPage } from "./pages/StatsPage";
+import { DownloadPage } from "./pages/DownloadPage";
 
 export function Routes() {
     return (
@@ -40,6 +41,9 @@ export function Routes() {
                     <LinkContainer to="/stats">
                         <NavItem>Statistics</NavItem>
                     </LinkContainer>
+                    <LinkContainer to="/download">
+                        <NavItem>Download</NavItem>
+                    </LinkContainer>
                     <LinkContainer to="/about">
                         <NavItem>About</NavItem>
                     </LinkContainer>
@@ -54,6 +58,7 @@ export function Routes() {
                 <Route path="/projects/:id/history" component={HistoryProject} />
                 <Route path="/projects/:id" component={Project} />
                 <Route path="/stats" component={StatsPage} />
+                <Route path="/download" component={DownloadPage}/>
                 <Route path="/about" component={About}/>
                 <ProtectedRoute path="/pending" exact component={PendingProjects} />
                 <ProtectedRoute path="/pending/:id" component={ReviewProject} />